OPE-12 split intake tool registration
